fix: modification de plusieurs code d'erreur à destination du client

+++ class/Src/Controller/Logout.php
@@ -9,10 +9,15 @@ class Logout extends \Src\Controller\Controller
   {
 
     if ($isApi) {
-      \Src\Api\Auth::protect();
+      if (!\Src\Api\Auth::protect()) {
+        http_response_code(403);
+        exit();
+      }
 
       \Src\Api\Auth::clearCookie("refresh_key");
-      return $this->deleteToken($id);
+      $this->deleteToken($id);
+
+      return \Src\App::sendApiData("ok");
     }
 
     \Src\Auth\Auth::protect();
